Added missing replay property and some refactoring
